services/bot/requirements.txt
python-telegram-bot==20.3
requests
websocket-client
python-dotenv

